Deployment improvements
Removed Makefile, because it was hard to explain whole deployment infrastructure and all the debugging issues.
Not using INCLUDE/EXCLUDE file lists, as those do not take long to copy between servers and is more readable.

Instead of Makefile trying to separate code, that need to be build on CI server (build.sh) and the one, than need to be executed on WEB server (deploy_playbook.yaml).

Generating ansible hosts file, because it was common to forget to update ansible user (the user/directory where to deploy file). This is intended to be generated (and overwritten in Jenkins CI)